fix: ui polish - dice bank display, safe color exploration, hand count
