+++ src/components/DiceBankDisplay.tsx
@@ -1,0 +1,75 @@
+import React from 'react'
+import { rollCounts } from 'common/attackRoll'
+import { DieValue } from 'common/die'
+import { ShellColor } from 'common/colors'
+
+const DIE_STYLE: Record<DieValue, { bg: string; text: string }> = {
+  B: { bg: '#222222', text: '#ffffff' },
+  W: { bg: '#eeeeee', text: '#222222' },
+  G: { bg: '#888888', text: '#ffffff' },
+  L: { bg: '#e8d44d', text: '#222222' },
+  T: { bg: '#e74c3c', text: '#ffffff' },
+}
+
+const WAVE_COLORS: Array<{ symbol: 'B' | 'W' | 'G'; color: ShellColor }> = [
+  { symbol: 'B', color: 'black' },
+  { symbol: 'W', color: 'white' },
+  { symbol: 'G', color: 'gray' },
+]
+
+const WAVE_DIE_COLORS = new Set<DieValue>(['B', 'W', 'G'])
+
+interface DiceBankDisplayProps {
+  bank: rollCounts
+  selectedColor?: ShellColor | null
+  onSelect?: (color: ShellColor) => void
+}
+
+export const DiceBankDisplay: React.FC<DiceBankDisplayProps> = ({
+  bank,
+  selectedColor = null,
+  onSelect,
+}) => {
+  const entries = (Object.entries(bank) as [DieValue, number][]).filter(
+    ([, count]) => count > 0,
+  )
+  if (entries.length === 0) return null
+  return (
+    <div style={{ display: 'flex', gap: 8, flexWrap: 'wrap', margin: '8px 0' }}>
+      {entries.map(([face, count]) => {
+        const s = DIE_STYLE[face]
+        const isSelectable = onSelect && WAVE_DIE_COLORS.has(face)
+        const color = WAVE_COLORS.find(c => c.symbol === face)?.color ?? null
+        const isSelected = color !== null && selectedColor === color
+        return (
+          <div
+            key={face}
+            style={{ display: 'flex', alignItems: 'center', gap: 4 }}>
+            <div
+              onClick={
+                isSelectable && color ? () => onSelect(color) : undefined
+              }
+              style={{
+                width: 32,
+                height: 32,
+                background: s.bg,
+                color: s.text,
+                border: isSelected ? '3px solid #2980b9' : '2px solid #555',
+                borderRadius: 6,
+                display: 'flex',
+                alignItems: 'center',
+                justifyContent: 'center',
+                fontWeight: 'bold',
+                fontSize: 14,
+                cursor: isSelectable ? 'pointer' : 'default',
+                boxShadow: isSelected ? '0 0 0 2px #2980b9' : 'none',
+              }}>
+              {face}
+            </div>
+            <span style={{ fontSize: 14, fontWeight: 'bold' }}>×{count}</span>
+          </div>
+        )
+      })}
+    </div>
+  )
+}
